cache config safety part 2
1) Mark config safe if the server reports the relevant config
   is up to date.
2) Mark config safe if we have been up for long enough to consider
   the device offline. This is currently 15s.
3) Simplify the retry_count file to be a single byte.
4) Break circular deps and various cleanups to make testing simpler.
5) (Unrelated) Remove the compression feature flag which we will
   never flip.
