+++ bd-api/Cargo.toml
@@ -37,4 +37,5 @@ assert_matches.workspace = true
 bd-stats-common.path     = "../bd-stats-common"
 bd-test-helpers          = { path = "../bd-test-helpers", default-features = false }
 ctor.workspace           = true
+mockall.workspace        = true
 tempfile.workspace       = true
