+++ bd-logger/src/builder.rs
@@ -238,7 +238,7 @@ impl LoggerBuilder {
       log.clone(),
     )?;
 
-    let updater = Box::new(client_config::Config::new(
+    let updater = Arc::new(client_config::Config::new(
       &self.params.sdk_directory,
       LoggerUpdate::new(
         buffer_manager.clone(),
@@ -263,10 +263,7 @@ impl LoggerBuilder {
       trigger_upload_tx,
       self.params.static_metadata,
       runtime_loader.clone(),
-      vec![
-        Box::new(bd_runtime::runtime::RuntimeManager::new(runtime_loader)),
-        updater,
-      ],
+      vec![runtime_loader, updater],
       Arc::new(SystemTimeProvider),
       network_quality_provider,
       log.clone(),
