+++ bd-logger/src/client_config_test.rs
@@ -7,6 +7,7 @@
 
 use super::{Config, Configuration};
 use anyhow::anyhow;
+use bd_client_common::{ConfigurationUpdate as _, HANDSHAKE_FLAG_CONFIG_UP_TO_DATE};
 use bd_proto::protos::client::api::configuration_update::{StateOfTheWorld, Update_type};
 use bd_proto::protos::client::api::ConfigurationUpdate;
 use bd_proto::protos::config::v1::config::BufferConfigList;
@@ -20,7 +21,7 @@ struct TestUpdate {
 
 #[async_trait::async_trait]
 impl super::ApplyConfig for TestUpdate {
-  async fn apply_configuration(&mut self, configuration: Configuration) -> anyhow::Result<()> {
+  async fn apply_configuration(&self, configuration: Configuration) -> anyhow::Result<()> {
     self
       .configuration_tx
       .send(configuration)
@@ -45,7 +46,7 @@ async fn process_and_load() {
   let (configuration_tx, mut configuration_rx) = channel(2);
   let update = TestUpdate { configuration_tx };
   let directory = TempDir::with_prefix("sdk").unwrap();
-  let mut config = Config::new(directory.path(), update);
+  let config = Config::new(directory.path(), update);
 
   config
     .process_configuration_update(&configuration_update())
@@ -57,9 +58,30 @@ async fn process_and_load() {
     configuration_rx.recv().await.unwrap()
   );
 
-  config.try_load_persisted_config_helper().await;
-
+  config.try_load_persisted_config().await;
   configuration_rx.recv().await;
+
+  // With no safety marking, we should have a retry count of 1.
+  assert_eq!(
+    std::fs::read(directory.path().join("config").join("retry_count")).unwrap(),
+    &[1]
+  );
+
+  // Getting a handshake without runtime being up to date should not do anything.
+  config.on_handshake_complete(0).await;
+  assert_eq!(
+    std::fs::read(directory.path().join("config").join("retry_count")).unwrap(),
+    &[1]
+  );
+
+  // Getting a handshake with runtime being up to date should mark the config as safe.
+  config
+    .on_handshake_complete(HANDSHAKE_FLAG_CONFIG_UP_TO_DATE)
+    .await;
+  assert_eq!(
+    std::fs::read(directory.path().join("config").join("retry_count")).unwrap(),
+    &[0]
+  );
 }
 
 #[tokio::test]
@@ -67,7 +89,7 @@ async fn cache_write_fails() {
   let (configuration_tx, mut configuration_rx) = channel(2);
   let update = TestUpdate { configuration_tx };
   let directory = TempDir::with_prefix("sdk").unwrap();
-  let mut config = Config::new(directory.path(), update);
+  let config = Config::new(directory.path(), update);
 
   // Create the config file, then make it read only. This should make the attempt to cache fail,
   // which should result in this being reported as an error.
@@ -97,7 +119,7 @@ async fn load_malformed_file() {
   let (configuration_tx, _configuration_rx) = channel(2);
   let update = TestUpdate { configuration_tx };
   let directory = TempDir::with_prefix("sdk").unwrap();
-  let mut config = Config::new(directory.path(), update);
+  let config = Config::new(directory.path(), update);
 
   let config_file = directory.path().join("config").join("protobuf.pb");
   std::fs::write(&config_file, "not a proto").unwrap();
