+++ bd-runtime/Cargo.toml
@@ -10,9 +10,11 @@ doctest = false
 
 [dependencies]
 anyhow.workspace      = true
+async-trait.workspace = true
 bd-client-common.path = "../bd-client-common"
 bd-proto.path         = "../bd-proto"
 log.workspace         = true
+parking_lot.workspace = true
 protobuf.workspace    = true
 time.workspace        = true
 tokio.workspace       = true
